requirements.txt
pandas >= 1.5.2
nose2 >= 0.12.0
openpyxl >= 3.0.10
black >= 22.12.0
flake8 >= 6.0.0
multipledispatch >= 0.6.0
pyyaml >= 5.3.1